requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
python-dotenv==1.0.0
bcrypt==4.1.1
dlib==19.24.2
